moved Resource object to core/mod to allow me to use it in save commands
